+++ CHANGELOG.md
@@ -2,7 +2,11 @@
 
 This file contains a centralizes a trace of all published crate versions, with their changes in short.
 
-## [multiversx-chain-vm-executor 0.1.0] - 2023-06-15
+## [klever-chain-vm-executor 0.2.0] - 2023-10-12
+- New VM hook: `managedGetBackTransfers`.
+- Memory fix.
+
+## [klever-chain-vm-executor 0.1.0] - 2023-06-15
 This is the initial official release of the VM executor interface. The purpose is for it to be used in the new smart contract debugger architecture.
 
 It targets VM 1.5 and integrates the Wasmer 2.2 implementation.